django-corsheaders requirements 추가
