+++ .requirements/production.txt
@@ -1,6 +1,8 @@
 boto3<1.7
 Django<2.1
 django-storages<1.7
+django-cors-headers<2.3
+djangorestframework<3.8
 psycopg2-binary<2.8
 Pillow<6.0.0
 raven<7
