+++ app/members/apis/auth.py
@@ -1,0 +1,28 @@
+from rest_framework.authtoken.models import Token
+from rest_framework.authtoken.serializers import AuthTokenSerializer
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+from ..serializers import UserSerializer
+
+
+class UserLoginAuthTokenAPIView(APIView):
+    def post(self, request):
+        serializer = AuthTokenSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
+        user = serializer.validated_data['user']
+        token, _ = Token.objects.get_or_create(user=user)
+        data = {
+            'token': token.key,
+            'user': UserSerializer(user).data,
+        }
+        return Response(data)
+
+
+class UserLogoutView(APIView):
+    def get(self, request):
+
+        data = {
+
+        }
+        return Response(data)
